+++ src/catalog/index.ts
@@ -1,5 +1,6 @@
 export * from "./config";
 export * from "./projects";
+export * from "./gtd";
 export * from "./mutations";
 export * from "./rename";
 export * from "./query";